refactoring, part 2
