+++ src/lib/inq/addInstruction.ts
