+++ src/lib/inq/cluster.ts
@@ -1,12 +1,12 @@
 import inquirer from "inquirer";
-import { DEFAULT_CLUSTER } from "../constants"
+import { CLUSTER } from "../constants"
 
 export default () => {
     const questions = [
         {
             name: 'cluster',
             type: 'input',
-            message: `Enter the rpc cluster to use (or enter for ${DEFAULT_CLUSTER}):`,
+            message: `Enter the rpc cluster to use (or enter for ${CLUSTER}):`,
           }
       ];
       return inquirer.prompt(questions);
