+++ src/lib/inq/settings.ts
